fix test by splitting module by features
