+++ src/test/java/it/pagopa/gov/rtdmsexporter/application/PagedDatabaseExportStepTest.java
@@ -1,9 +1,14 @@
 package it.pagopa.gov.rtdmsexporter.application;
 
 import com.mongodb.MongoException;
-import it.pagopa.gov.rtdmsexporter.configuration.AppConfiguration;
-import it.pagopa.gov.rtdmsexporter.configuration.ExportJobModule;
-import it.pagopa.gov.rtdmsexporter.configuration.MockMongoConfiguration;
+import io.reactivex.rxjava3.core.Scheduler;
+import io.reactivex.rxjava3.schedulers.Schedulers;
+import it.pagopa.gov.rtdmsexporter.application.acquirer.SaveAcquirerFileStep;
+import it.pagopa.gov.rtdmsexporter.application.paymentinstrument.NewExportedSubscriber;
+import it.pagopa.gov.rtdmsexporter.configuration.AcquirerModule;
+import it.pagopa.gov.rtdmsexporter.configuration.DatabaseMockConfiguration;
+import it.pagopa.gov.rtdmsexporter.configuration.DatabaseReaderModule;
+import it.pagopa.gov.rtdmsexporter.domain.acquirer.AcquirerFileRepository;
 import it.pagopa.gov.rtdmsexporter.infrastructure.mongo.CardEntity;
 import it.pagopa.gov.rtdmsexporter.utils.HashStream;
 import org.junit.jupiter.api.AfterEach;
@@ -12,6 +17,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Query;
@@ -24,10 +32,11 @@ import org.springframework.test.context.support.DependencyInjectionTestExecution
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static it.pagopa.gov.rtdmsexporter.configuration.ExportJobModule.ACQUIRER_GENERATED_FILE;
+import static it.pagopa.gov.rtdmsexporter.configuration.AcquirerModule.ACQUIRER_GENERATED_FILE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertThrows;
 import static org.mockito.ArgumentMatchers.*;
@@ -36,8 +45,7 @@ import static org.mockito.Mockito.when;
 
 @ExtendWith(SpringExtension.class)
 @ExtendWith(MockitoExtension.class)
-@Import(MockMongoConfiguration.class)
-@ContextConfiguration(classes = {ExportJobModule.class, AppConfiguration.class})
+@ContextConfiguration(classes = { PagedDatabaseExportStepTest.MockDeps.class, AcquirerModule.class, DatabaseReaderModule.class })
 @TestExecutionListeners({DependencyInjectionTestExecutionListener.class})
 @TestPropertySource(locations = "classpath:application.yml")
 class PagedDatabaseExportStepTest {
@@ -48,8 +56,12 @@ class PagedDatabaseExportStepTest {
   @Autowired
   private MongoTemplate mongoTemplate;
 
+  @Autowired
+  private NewExportedSubscriber newExportedSubscriber;
+
   @BeforeEach
   void setup() {
+    when(newExportedSubscriber.apply(any())).thenReturn(CompletableFuture.completedFuture(0L));
   }
 
   @AfterEach
@@ -80,5 +92,21 @@ class PagedDatabaseExportStepTest {
   void whenMongoThrowsExceptionThenAnExceptionIsThrown() {
     when(mongoTemplate.find(any(Query.class), eq(CardEntity.class), anyString())).thenThrow(new MongoException("Unexpected error"));
     assertThrows(RuntimeException.class, () -> pagedDatabaseExportStep.execute());
+  }
+
+  @TestConfiguration
+  @Import(DatabaseMockConfiguration.class)
+  static class MockDeps {
+    @MockBean
+    SaveAcquirerFileStep saveAcquirerFileStep;
+    @MockBean
+    AcquirerFileRepository acquirerFileRepository;
+    @MockBean
+    NewExportedSubscriber newExportedSubscriber;
+
+    @Bean
+    Scheduler rxScheduler() {
+      return Schedulers.io();
+    }
   }
 }
